# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Channel:
    properties:
      created_at:
        type: string
      created_by:
        type: integer
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      settings:
        $ref: '#/definitions/models.ChannelSettings'
      tenant_id:
        type: integer
      type:
        $ref: '#/definitions/models.ChannelType'
      updated_at:
        type: string
    type: object
  models.ChannelSettings:
    properties:
      allow_bots:
        type: boolean
      allow_files:
        type: boolean
      allowed_file_types:
        items:
          type: string
        type: array
      max_file_size:
        type: integer
      notifications:
        type: boolean
      retention_days:
        type: integer
    type: object
  models.ChannelType:
    enum:
    - public
    - private
    - direct
    type: string
    x-enum-varnames:
    - ChannelTypePublic
    - ChannelTypePrivate
    - ChannelTypeDirect
...
  models.Message:
    properties:
      content:
        type: string
      group_id:
        type: integer
      group_name:
        type: string
      receiver:
        type: string
      receiver_id:
        type: integer
      sender:
        type: string
      sender_id:
        type: integer
      tenant_id:
        type: integer
      timestamp:
        type: string
      topic_name:
        type: string
      type:
        description: '"private", "group", "notification"'
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A chat system API with real-time messaging, file sharing, and support
    ticket management.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Chat System API
  version: "1.0"
paths:
...
  /api/v1/channels:
    post:
      consumes:
      - application/json
      description: Create a new chat channel
      parameters:
      - description: Channel details
        in: body
        name: channel
        required: true
        schema:
          $ref: '#/definitions/models.Channel'
      produces:
      - application/json
      responses:
        "201":
          description: Channel created successfully
          schema:
            $ref: '#/definitions/models.Channel'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new channel
      tags:
      - Chat
...
  /api/v1/channels/{channel_id}/messages:
    get:
      consumes:
      - application/json
      description: Get messages from a specific channel with pagination
      parameters:
      - description: Channel ID
        in: path
        name: channel_id
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of messages per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of messages
          schema:
            items:
              $ref: '#/definitions/models.Message'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Channel not found
          schema:
            additionalProperties: true
            type: object
      summary: Get channel messages
      tags:
      - Chat
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...